# api/openapi.yaml
openapi: 3.0.1
info:
  title: MERN   API
  version: 1.0.0
  description: API endpoints [social media app, classroom app]
  contact:
    email: fmg3ckali@gmail.com
    name: Nathaniel Cobbinah
    url: https://shy-lime-dibbler-garb.cyclic.app/
  license:
    name: MIT
    url: https://choosealicense.com/licenses/mit/
paths:
...
  /api/enrollment/enrolled:
    get:
      summary:  returns list of enrollments for a specific student using their id when they login successfully
      description: ''
      operationId: listEnrolled
      responses:
        '200':
          description: Default sample response
          content:
            application/json:
              examples:
                 returns list of enrollments for a specific student using their id when they login successfully:
                  value: |-
                    [
                      {
                        "course":{
                          "id": "xxxxx",
                          "name":"yyyyyy",
                          "category":"pppppp",
                        },
                        "student": "idxxxwey",
                        "lessonStatus":[
                          {
                            "lesson":{
                              "title":"zzzz",
                              "content":"mmmm",
                              "resource_url":"url..."
                            }
                          },
                          complete: false
                        ],
                        "enrolled": Date(),
                        "updated": Date(),
                        "completed": Date()
                      }
                    ]
        default:
          description: Default error sample response
      tags:
        - Enrollment 
      security:
        - bearerAuth: []
  /api/enrollment/stats/{courseId}:
    get:
      summary: Queries the Enrollments collection in the database to calculate the stats for a specific course
      description: ''
      operationId: enrollmentStats
      parameters:
        - name: courseId
          in: path
          description: Queries the Enrollments collection in the database to calculate the stats for a specific course
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        '200':
          description: Default sample response
          content:
            application/json:
              examples:
                return a stats object containing the total enrollments and total completions for the course, as identified by the courseId :
                  value: |-
                    "stats": { 
                      "totalEnrolled":7777,
                      "totalCompleted": 99878,
                    }
        default:
          description: Default error sample response
      tags:
        - Enrollment   
  /api/enrollment/new/{courseId}:
    get:
      summary: Creates a new enrollment for a user, when they choose to enroll in a course
      description: ''
      operationId: createEnrollment
      parameters:
        - name: courseId
          in: path
          description: Used to retrieve the corresponding course from the database
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        '200':
          description: Default sample response
          content:
            application/json:
              examples:
                returns an enrollment object with course, user and lessons  :
                  value: |-
                    { 
                      "course":{
                      },
                      "student": "ObjectId(i768778687)",
                      "lessonStatus":[
                         "lesson":{
                           "title": "ououoiu",
                           "content": "sljdsdflsjd",
                           "resource_url":"url()"
                         },
                         "complete":boolean
                      ],
                      "enrolled":Date(),
                      "updated":Date(),
                      "completed":Date()
                    }
        default:
          description: Default error sample response
      tags:
        - Enrollment 
      security:
        - bearerAuth: []
  /api/enrollment/{enrollmentId}:
    get:
      summary:  Returns the enrollment details for a given user who is enrolled to a course
      description: ''
      operationId: readEnrollments
      parameters:
        - name: enrollmentId
          in: path
          description: The enrollmentID is used to query the Enrollments collection by the provided ID, and if a matching enrollment document is found, we ensure that the referenced course, the nested course instructor, and the referenced student details are also populated and returned from the database. 
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        '200':
          description: Default sample response
          content:
            application/json:
              examples:
                returns an course enrolled by a user:
                  value: |-
                    { 
                      "course":{
                      },
                      "student": "ObjectId(i768778687)",
                      "lessonStatus":[
                         "lesson":{
                           "title": "ououoiu",
                           "content": "sljdsdflsjd",
                           "resource_url":"url()"
                         },
                         "complete":boolean
                      ],
                      "enrolled":Date(),
                      "updated":Date(),
                      "completed":Date()
                    }
        default:
          description: Default error sample response
      tags:
        - Enrollment
      security:
        - bearerAuth: []
...
  /auth/signout:
    summary: Endpoint to logout user from the system
    description: Clears userToken, thus logging out user from the system
    get:
      summary: Logout user from the system
      description: >-
        Clear validToken, and disable access to userInfo unless user is loggedIn
        again
      operationId: signout
      responses:
        '200':
          content:
            application/json:
              examples:
                User signed out successfully:
                  value: signed out
          description: Clears userToken from the system
        default:
          description: Default error sample response
      tags:
        - User Authentication
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT